Backend/requirements.txt
\
langchain-text-splitters
mem0ai
# Additional dependencies might be needed from langchain_arch/requirements.txt?
# Let's assume the needed langchain deps are covered here.
# We need an HTTP client. Let's add httpx.
httpx>=0.27.0 # For making async requests from tools

# --- Other internal packages --- #
# Assuming langchain_arch might be installed as editable or handled via PYTHONPATH
